Update AppQuery to relay 11
