+++ package.json
@@ -62,14 +62,14 @@
     "@types/react-native": "^0.62.13",
     "@types/react-native-dotenv": "^0.2.0",
     "@types/react-native-video": "^5.0.3",
-    "@types/react-relay": "^7.0.8",
+    "@types/react-relay": "^11.0.1",
     "@types/react-test-renderer": "^16.9.2",
-    "@types/relay-runtime": "^9.1.5",
+    "@types/relay-runtime": "^10.1.10",
     "@typescript-eslint/eslint-plugin": "^4.6.0",
     "@typescript-eslint/parser": "^4.6.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^24.9.0",
-    "babel-plugin-relay": "^9.1.0",
+    "babel-plugin-relay": "^11.0.1",
     "chai": "^4.2.0",
     "dotenv": "^8.2.0",
     "eslint": "^7.12.0",
@@ -96,8 +96,9 @@
     "prettier": "^2.1.2",
     "qrcode-terminal": "^0.12.0",
     "react-test-renderer": "16.11.0",
-    "relay-compiler": "^9.1.0",
+    "relay-compiler": "^11.0.1",
     "relay-compiler-language-typescript": "^12.0.3",
+    "relay-config": "^11.0.1",
     "relay-devtools": "^1.4.0",
     "ts-node": "^8.6.2",
     "typescript": "^3.8.3"
